CFY-3 add teardown; fix test bug

+++ cosmo_tester/framework/cloud_terminator.py
@@ -18,21 +18,24 @@ __author__ = 'nirb'
 import yaml
 import novaclient.v1_1.client as nvclient
 import neutronclient.v2_0.client as neclient
+from path import path
 
-from cosmo_tester.framework.util import get_resource_path
 
+def teardown(cloud_config_file_path,
+             remove_server=True,
+             remove_network=False,
+             remove_security_groups=False,
+             remove_keypairs=False):
 
-def teardown(cloud_config_file_name):
-    cloud_config_file_path = get_resource_path(cloud_config_file_name)
-    creds = get_client_creds(cloud_config_file_path)
+    cloud_config = yaml.load(path(cloud_config_file_path).text())
+
+    creds = get_client_creds(cloud_config)
     nova = nvclient.Client(**creds)
     neutron = neclient.Client(username=creds['username'],
                               password=creds['api_key'],
                               tenant_name=creds['project_id'],
                               region_name=creds['region_name'],
                               auth_url=creds['auth_url'])
-
-    cloud_config = yaml.load(open(cloud_config_file_path, 'r'))
 
     networking = cloud_config['networking']
     compute = cloud_config['compute']
@@ -45,35 +48,38 @@ def teardown(cloud_config_file_name):
     mng_security_group_name = networking['management_security_group']['name']
     agents_security_group_name = networking['agents_security_group']['name']
 
-    for server in nova.servers.list():
-        if mng_instance_name == server.human_id:
-            server.delete()
-            break
+    if remove_server:
+        for server in nova.servers.list():
+            if mng_instance_name == server.human_id:
+                server.delete()
+                break
 
-    for network in neutron.list_networks()['networks']:
-        if network_name == network['name']:
-            neutron.delete_network(network['id'])
-            break
+    if remove_network:
+        for network in neutron.list_networks()['networks']:
+            if network_name == network['name']:
+                neutron.delete_network(network['id'])
+                break
 
-    for keypair in nova.keypairs.list():
-        if mng_keypair_name == keypair.name:
-            nova.keypairs.delete(keypair)
-        elif agents_keypair_name == keypair.name:
-            nova.keypairs.delete(keypair)
+    if remove_keypairs:
+        for keypair in nova.keypairs.list():
+            if mng_keypair_name == keypair.name:
+                nova.keypairs.delete(keypair)
+            elif agents_keypair_name == keypair.name:
+                nova.keypairs.delete(keypair)
 
-    for sg in nova.security_groups.list():
-        if mng_security_group_name == sg.name:
-            nova.security_groups.delete(sg)
-        elif agents_security_group_name == sg.name:
-            nova.security_groups.delete(sg)
+    if remove_security_groups:
+        for sg in nova.security_groups.list():
+            if mng_security_group_name == sg.name:
+                nova.security_groups.delete(sg)
+            elif agents_security_group_name == sg.name:
+                nova.security_groups.delete(sg)
 
 
-def get_client_creds(cloud_config_file_path):
-    config_yaml = yaml.load(open(cloud_config_file_path, 'r'))
+def get_client_creds(cloud_config):
     return {
-        'username': config_yaml['keystone']['username'],
-        'api_key': config_yaml['keystone']['password'],
-        'auth_url': config_yaml['keystone']['auth_url'],
-        'project_id': config_yaml['keystone']['tenant_name'],
-        'region_name': config_yaml['compute']['region']
+        'username': cloud_config['keystone']['username'],
+        'api_key': cloud_config['keystone']['password'],
+        'auth_url': cloud_config['keystone']['auth_url'],
+        'project_id': cloud_config['keystone']['tenant_name'],
+        'region_name': cloud_config['compute']['region']
     }
